Revamp project structure and refactor code
