+++ negate/__init__.py
@@ -1,6 +1,5 @@
 from .negate import Negator
-from .tokens import Token
+from .utils.tokens import Token
 
 # Don't expose the following submodules.
-#del globals()["negate"]
-#del globals()["tokens"]
+del globals()["negate"]
