+++ negate/negators/de/__init__.py
@@ -1,0 +1,1 @@
+LANG_NAME: str = "German"
