+++ negate/negators/de/negator.py
@@ -1,24 +1,34 @@
-from negate.negator import Negator_ABC
-from typing import Dict, Optional, Union
+"""German Negation."""
+
 import importlib
+import logging
 import os
 import sys
 from contextlib import contextmanager
+from typing import Any, Dict, List, Optional, Set, Union
 
 import spacy
 from DERBI.derbi import DERBI
+from negate.base import BaseNegator
+from negate.utils.tokens import Token
 from spacy.lang.de import German
+from spacy.symbols import AUX, VERB
 from spacy.tokens import Doc as SpacyDoc
 from spacy.tokens import Token as SpacyToken
 
-from negate.tokens import Token
 
+class Negator(BaseNegator):
+    """Negator for the German language."""
 
-class Negator_DE(Negator_ABC):
-
-    def __init__(self, use_transformers: Optional[bool] = None, use_gpu: Optional[bool] = None,
-                 fail_on_unsupported: Optional[bool] = None, log_level: Optional[int] = None):
-        """Instanciate a :obj:`Negator`.
+    def __init__(
+        self,
+        use_transformers: Optional[bool] = None,
+        use_gpu: Optional[bool] = None,
+        fail_on_unsupported: Optional[bool] = None,
+        log_level: Optional[int] = None,
+        **kwargs,
+    ):
+        """Instanciate a German Negator.
 
         Args:
             use_transformers (:obj:`Optional[bool]`, defaults to :obj:`False`):
@@ -43,7 +53,33 @@ class Negator_DE(Negator_ABC):
             :obj:`RuntimeError`: If the sentence is not supported and
             :arg:`fail_on_unsupported` is set to :obj:`True`.
         """
-        super().__init__(use_transformers, use_gpu, fail_on_unsupported, log_level)
+        if use_transformers is None:
+            use_transformers = False
+        if use_gpu is None:
+            use_gpu = False
+        if fail_on_unsupported is None:
+            fail_on_unsupported = False
+        if log_level is None:
+            log_level = logging.INFO
+
+        # Set up logger.
+        logging.basicConfig(
+            format="%(levelname)s: %(message)s",
+            level=log_level
+        )
+        self.logger = logging.getLogger(__class__.__name__)
+        self.fail_on_unsupported = fail_on_unsupported
+        # Load spaCy model. If not available locally, the model will be first
+        # installed.
+        if use_transformers and use_gpu:
+            spacy.require_gpu()
+        else:
+            spacy.require_cpu()
+        self.spacy_model = self._initialize_spacy_model(use_transformers)
+        # Store whether tokens have a whitespace after them. This is used later
+        # on for de-tokenization.
+        SpacyToken.set_extension("has_space_after", default=True, force=True)
+
         self.derbi = DERBI(self.spacy_model)
         self._de_initialize_nicht_replace_table()
         self._de_initialize_kein_table()
@@ -55,13 +91,37 @@ class Negator_DE(Negator_ABC):
         self._de_initialize_reverse_two_part_phrase_table()
 
     def negate_sentence(
-            self,
-            sentence: str,
-            strategy: tuple = ("kein", "nicht", "phrase"),
-    ) -> set[str]:
+        self,
+        sentence: str,
+        **kwargs: Dict[str, Any],
+    ) -> List[str]:
+        """Negate a sentence.
+
+        Affirmative sentences will be turned into negative ones and vice versa.
+
+        .. note::
+
+            Currently, only sentences that contain at least one verb are
+            supported. The output of non-supported sentences might be arbitrary.
+
+        Args:
+            sentence (:obj:`str`):
+                The sentence to negate.
+            strategy (:obj:`Optional[List[str]]`, defaults to ``["kein", "nicht", "phrase"]``):
+                The negation strategy to use, i.e., whether to negate by adding
+                the "kein" or "nicht" negation particles, or perform a phrase
+                negation.
+
+        Returns:
+            :obj:`List[str]`: The negated sentence(s).
+        """
         results = set()
         if not sentence:
-            return set()
+            return []
+
+        strategy = kwargs.get("strategy")
+        if strategy is None:
+            strategy = ["kein", "nicht", "phrase"]
 
         for contraction in self._contraction_table:
             sentence = sentence.replace(contraction, self._contraction_table[contraction])
@@ -71,11 +131,11 @@ class Negator_DE(Negator_ABC):
 
         if not root:
             self._handle_unsupported()
-            return set()
+            return []
 
         un_negated = self._un_negate_sentence(sentence, doc)
         if un_negated:
-            return un_negated
+            return list(un_negated)
 
         # Edge case "weder noch"
         exception_list = [" weder ", "Weder ", " sowohl ", "Sowohl"]
@@ -97,9 +157,256 @@ class Negator_DE(Negator_ABC):
                 final_result = final_result.replace(contraction, self._reverse_contraction_table[contraction])
             final_results.add(final_result)
 
-        return final_results
+        return list(final_results)
 
-    def _un_negate_sentence(self, sentence: str, doc: SpacyDoc) -> set[str]:
+    def _parse(self, string_: str) -> SpacyDoc:
+        """Parse a string.
+
+        This method cleans up the string and tokenizes it. The resulting
+        :obj:`SpacyDoc` object, also includes information about whitespaces to
+        facilitate de-tokenization later on.
+
+        Args:
+            string_ (:obj:`str`):
+                The string to parse.
+
+        Returns:
+            :obj:`SpacyDoc`: The string tokenized into a spaCy document.
+        """
+        # Remove extra whitespaces and other non-printable chars.
+        string_ = self._remove_extra_whitespaces(string_)
+        # Tokenize.
+        doc = self.spacy_model(string_)
+        i = 0  # Used to determine whitespaces.
+        for tk in doc:
+            has_space_after: bool = (
+                i+len(tk) < len(string_) and (string_[i+len(tk)] == " ")
+            )
+            tk._.has_space_after = has_space_after
+            i += len(tk) + int(has_space_after)
+        return doc
+
+    def _get_entry_point(
+        self,
+        doc: SpacyDoc,
+        contains_inversion: bool
+    ) -> Optional[SpacyToken]:
+        """Choose a suitable verb to attempt negating first, if any.
+
+        Args:
+            doc (:obj:`SpacyDoc`):
+                The spaCy document in which to find the entry point.
+            contains_inversion (:obj:`bool`):
+                Whether the sentence contains an inversion or not.
+
+        Returns:
+            :obj:`Optional[SpacyToken]`: The chosen entry point (verb), or
+            :obj:`None` if the sentence has no root, or contains no verbs.
+        """
+        if contains_inversion:
+            entry_point = [tk for tk in doc
+                           if self._is_aux(tk) or self._is_verb(tk)]
+            if entry_point:
+                return entry_point[0]
+        root = self._get_root(doc)
+        if root is None:  # nothing we can do
+            return None
+        # If the root token is not an AUX or a VERB, look for an AUX or
+        # VERB in its children.
+        if not (self._is_aux(root) or self._is_verb(root)):
+            entry_point = None
+            if root.children:
+                entry_point = [tk for tk in root.children
+                                if self._is_aux(tk) or self._is_verb(tk)]
+            # No AUX or VERB found in the root children -> Take the first
+            # AUX or VERB in the sentence, if any.
+            if not entry_point:
+                entry_point = [tk for tk in doc
+                                if self._is_aux(tk) or self._is_verb(tk)]
+            return entry_point[0] if entry_point else None
+        return root
+
+    def _get_root(self, doc: SpacyDoc) -> Optional[SpacyToken]:
+        """Get the root token in a spaCy document, if any.
+
+        Args:
+            doc (:obj:`SpacyDoc`):
+                The spaCy document to get the root from.
+
+        Returns:
+            :obj:`Optional[SpacyToken]`: The root token, or :obj:`None` if the
+            sentence has no root.
+        """
+        root = [tk for tk in doc if tk.dep_ == "ROOT"]
+        return root[0] if root else None
+
+    def _capitalize_first_letter(self, string_: str) -> str:
+        """Uppercase the first letter of a string.
+
+        The capitalization of the rest of the string remains unchanged.
+
+        Args:
+            string_ (:obj:`str`):
+                The string whose first letter to uppercase.
+
+        Returns:
+            :obj:`str`: The string with its first letter uppercased.
+        """
+        if not string_:
+            return ""
+        return f"{string_[0].upper()}{string_[1:]}"
+
+    def _remove_extra_whitespaces(self, string_: str) -> str:
+        """Remove any duplicated whitespaces in a string.
+
+        Args:
+            string_ (:obj:`str`):
+                The string in which to remove any extra whitespaces.
+
+        Returns:
+            :obj:`str`: The string with one whitespace at most between words.
+        """
+        if not string_:
+            return ""
+        return " ".join(string_.split())
+
+    def _is_aux(self, token: SpacyToken) -> bool:
+        """Determine whether a token is an auxiliary verb.
+
+        Args:
+            token (:obj:`SpacyToken`):
+                The spaCy token to determine whether it is auxiliary.
+
+        Returns:
+            :obj:`bool`: :obj:`True` if the token is an auxiliary verb,
+            otherwise :obj:`False`.
+        """
+        if not token:
+            return False
+        return token.pos == AUX
+
+    def _is_verb(self, token: SpacyToken) -> bool:
+        """Determine whether a token is a non-auxiliary verb.
+
+        .. note::
+
+            If you want to check if a token is either an auxiliary *or* a verb,
+            you can use this method in combination with :meth:`Negator._is_aux`.
+
+        Args:
+            token (:obj:`SpacyToken`):
+                The spaCy token to determine whether it is a non-auxiliary verb.
+
+        Returns:
+            :obj:`bool`: :obj:`True` if the token is a non-auxiliary verb,
+            otherwise :obj:`False`.
+        """
+        if not token:
+            return False
+        return token.pos == VERB
+
+    @staticmethod
+    def _is_full_sentence(root: SpacyToken) -> bool:
+        """Check if it is a full sentence"""
+        subject = [x for x in root.children if x.dep_ == "sb"]
+        if not subject:
+            return False
+        return subject[0].i < root.i
+
+    @staticmethod
+    def _find_verb_form(token: SpacyToken) -> str:
+        """find the verb form of token"""
+        result = ""
+        morph_dict = token.morph.to_dict()
+        if "VerbForm" in morph_dict:
+            result = morph_dict["VerbForm"]
+        return result
+
+    @staticmethod
+    def _find_definite(token: SpacyToken) -> str:
+        """find the definite type of token i.e. definite or indefinite"""
+        result = ""
+        morph_dict = token.morph.to_dict()
+        if "Definite" in morph_dict:
+            result = morph_dict["Definite"]
+        return result
+
+    @staticmethod
+    def _find_number(token: SpacyToken) -> str:
+        """find the number type of token i.e. plural or singular"""
+        result = ""
+        morph_dict = token.morph.to_dict()
+        if "Number" in morph_dict:
+            result = morph_dict["Number"]
+        return result
+
+    @staticmethod
+    def _find_case(token: SpacyToken) -> str:
+        """find the case type of token e.g. nominative, accusative, etc."""
+        result = ""
+        morph_dict = token.morph.to_dict()
+        if "Case" in morph_dict:
+            result = morph_dict["Case"]
+        return result
+
+    @staticmethod
+    def _find_gender(token: SpacyToken) -> str:
+        """find the gender type of token i.e. feminine, masculine, neutral."""
+        result = ""
+        morph_dict = token.morph.to_dict()
+        if "Gender" in morph_dict:
+            result = morph_dict["Gender"]
+        return result
+
+    @staticmethod
+    def _find_last_word(doc: SpacyDoc) -> SpacyToken:
+        """find the last word"""
+        if doc[-1].pos_ == "PUNCT":
+            return doc[-2]
+        return doc[-1]
+
+    def _compile_sentence(
+        self,
+        doc: SpacyDoc,
+        remove_tokens: Optional[List[int]] = None,
+        add_tokens: Optional[Dict[int, Token]] = None
+    ) -> str:
+        """Process and de-tokenize a spaCy document back into a string.
+
+        Args:
+            doc (:obj:`SpacyDoc`):
+                The spaCy document.
+            remove_tokens (:obj:`Optional[List[int]]`):
+                The indexes of the tokens to remove from the document, if any.
+            add_tokens (:obj:`Optional[Dict[int, Token]]`):
+                The tokens to add to the document, if any. These are specified
+                as a dictionary whose keys are the indexes in which to insert
+                the new tokens, which are the respective values.
+
+        Returns:
+            :obj:`str`: The resulting, de-tokenized string including the
+            removal/addition of tokens, if any.
+        """
+        if remove_tokens is None:
+            remove_tokens = []
+        if add_tokens is None:
+            add_tokens = {}
+        else:
+            add_tokens = dict(sorted(add_tokens.items()))  # sort by index
+        tokens = [Token(tk.text, tk._.has_space_after) for tk in doc]
+        for i in remove_tokens:
+            tokens[i] = Token(text="", has_space_after=False)
+        for count, item in enumerate(add_tokens.items()):
+            i, tk = item
+            tokens.insert(i+count, tk)
+        return self._capitalize_first_letter(
+            self._remove_extra_whitespaces(
+                "".join([f"{tk.text}{' '*int(tk.has_space_after)}"
+                         for tk in tokens])
+                )
+            )
+
+    def _un_negate_sentence(self, sentence: str, doc: SpacyDoc) -> Set[str]:
         # un negate phrases
         un_negated_sentences_phrases = self._reverse_negate_phrases(sentence)
         if un_negated_sentences_phrases:
@@ -143,7 +450,7 @@ class Negator_DE(Negator_ABC):
     def _negate_kein(
             self,
             doc: SpacyDoc
-    ) -> set[str]:
+    ) -> Set[str]:
         results = set()
         kein_noun_dicts = [self._generate_kein_dict(x) for x in doc if x.pos_ == "NOUN"]
         kein_noun_dicts = [x for x in kein_noun_dicts if x]
@@ -162,7 +469,7 @@ class Negator_DE(Negator_ABC):
             self,
             doc: SpacyDoc,
             root: Union[Token, SpacyToken]
-    ) -> set[str]:
+    ) -> Set[str]:
         results = set()
         svps = any(x.dep_ == "svp" for x in doc)
         adps = any(x.pos_ == "ADP" for x in doc)
@@ -187,7 +494,7 @@ class Negator_DE(Negator_ABC):
 
         return results
 
-    def _negate_phrases(self, sentence: str) -> set[str]:
+    def _negate_phrases(self, sentence: str) -> Set[str]:
         results = set()
 
         for phrase in self._phrase_table.keys():
@@ -205,7 +512,7 @@ class Negator_DE(Negator_ABC):
 
         return results
 
-    def _reverse_negate_phrases(self, sentence: str) -> set[str]:
+    def _reverse_negate_phrases(self, sentence: str) -> Set[str]:
 
         for phrase in self._reverse_phrase_table.keys():
             if phrase in sentence:
@@ -284,7 +591,7 @@ class Negator_DE(Negator_ABC):
             root: Union[Token, SpacyToken],
             doc: SpacyDoc,
             dont_negate_at_end: bool,  # noqa FBT001
-    ) -> set[str]:
+    ) -> Set[str]:
         adpositions = [x for x in doc if x.pos_ == "ADP"]
         results = set()
         for adposition in adpositions:
@@ -319,7 +626,7 @@ class Negator_DE(Negator_ABC):
     ) -> str:
         return self._negate_before_token(self._find_last_word(doc), doc)
 
-    def _generate_kein_dict(self, noun_token: Union[Token, SpacyToken]) -> dict | bool:
+    def _generate_kein_dict(self, noun_token: Union[Token, SpacyToken]) -> Union[bool, Dict]:
         children = list(noun_token.children)
         # No children to the right
         children = [child for child in children if child.i < noun_token.i]
