+++ negate/negators/en/__init__.py
@@ -1,0 +1,1 @@
+LANG_NAME: str = "English"
