+++ negate/version.py
@@ -3,6 +3,6 @@
 # Negate version.
 __version__ = "1.1.3"
 
-# spaCy models version.
+# spaCy models versions.
 EN_CORE_WEB_MD_VERSION: str = "3.7.0"
 EN_CORE_WEB_TRF_VERSION: str = "3.7.2"
