+++ tests/conftest.py
@@ -24,11 +24,15 @@ def pytest_generate_tests(metafunc):
             if not metafunc.config.getoption("use_cpu"):
                 with suppress(ValueError, NotImplementedError):
                     # `use_gpu` ignored if `use_transformers` is False.
-                    negator_model = Negator(use_transformers=use_transformers,
+                    negator_model = Negator(language="en",
+                                            use_transformers=use_transformers,
                                             use_gpu=True)
                     # If GPU is unsupported, we fallback to CPU.
                     negator_model.negate_sentence("I will now check GPU support!")
             else:
                 negator_model = Negator(
-                    use_transformers=use_transformers, use_gpu=False)
+                    language="en",
+                    use_transformers=use_transformers,
+                    use_gpu=False
+                )
         metafunc.parametrize("negator", [negator_model])
